+++ src/app/routes/actions.ts
@@ -3,8 +3,7 @@
  */
 import * as pecorino from '@pecorino/domain';
 import { Router } from 'express';
-// tslint:disable-next-line:no-submodule-imports
-import { query } from 'express-validator/check';
+import { query } from 'express-validator';
 import * as mongoose from 'mongoose';
 
 import authentication from '../middlewares/authentication';