+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 口座検索条件拡張
 - 口座開設を複数口座に対応
 - update mongoose
+- update express-validator
 
 ### Deprecated